refactor(tests): consolidate BotTurnTrigger tests into ai/ directory
Merged duplicate test file from src/server/__tests__/BotTurnTrigger.test.ts
into src/server/__tests__/ai/BotTurnTrigger.test.ts. Also updated the
advanceTurn assertion to pass expectedCurrentIndex (review feedback fix).
